+++ src/app/interface/observation.interface.ts
@@ -1,5 +1,6 @@
 import { ObservationCategory } from "./observationcategory.interface";
 import { User } from "./user.interface";
+
 export interface Observation {
   obs_id: number;
   description: string;
@@ -13,3 +14,44 @@ export interface Observation {
   usr_id_obs_mod: number;
   observationCategory: ObservationCategory;
 }
+  export interface ObservacionDTO{
+     usr2_email: String ;
+     usr2_ap_pat: String
+     usr2_nom: String ;
+     usr2_id: number;
+     obs_id: number;
+    apr_ger:String;
+    operationalApproval: String;
+    technicalApproval:String;
+    description:String;
+    creationDate:Date;
+    modificationDate:Date;
+    usr1_id: number;
+    usr_id_obs:number;
+    usr_id_obs_mod: number;
+    id: number;
+    name:String;
+    firstLastname:String;
+    email:String;
+  }
+  export interface CatObservacionDTO{
+     usr2_email: String ;
+     usr2_ap_pat: String
+     usr2_nom: String ;
+     usr2_id: number;
+     id:number;
+     cat_description:string;
+     obs_id: number;
+    apr_ger:String;
+    operationalApproval: String;
+    technicalApproval:String;
+    description:String;
+    creationDate:Date;
+    modificationDate:Date;
+    usr1_id: number;
+    usr_id_obs:number; 
+    usr_id_obs_mod: number ;
+    name:String;
+    firstLastname:String;
+    email:String;
+  }
